requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.2.1
backcall==0.2.0
Bottleneck==1.3.5
certifi==2022.12.7
comm==0.1.2
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
Django==3.2
django-cors-headers==3.13.0
django-dotenv==1.4.2
djangorestframework==3.14.0
entrypoints==0.4
executing==1.2.0
fonttools==4.25.0
h5py==3.6.0
hic-straw==1.3.1
ipykernel==6.19.4
ipyparallel==8.4.1
ipython==8.8.0
jedi==0.18.2
jupyter_client==7.4.8
jupyter_core==5.1.2
kiwisolver==1.4.4
matplotlib==3.6.2
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.23.4
packaging==22.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==22.2.2
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.10.1
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
pyzmq==24.0.1
scipy==1.9.3
seaborn==0.11.2
setuptools==65.5.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
wcwidth==0.2.5
wheel==0.37.1
